IOS-4956 Add toggle

+++ Anytype/Sources/PresentationLayer/Modules/Chat/ChatViewModel.swift
@@ -130,6 +130,7 @@ final class ChatViewModel: ObservableObject, MessageModuleOutput, ChatActionProv
     }
     
     func onAppear() {
+        guard FeatureFlags.readMessagesFromNotificationsCenterOnChatAppear else { return }
         notificationsCenterService.removeDeliveredNotifications(for: [chatId])
     }
